Feature/pricing

+++ src/main/java/com/ahlenius/wigell_cinema/dto/bookingDto/BookingResponse.java
@@ -1,6 +1,7 @@
 package com.ahlenius.wigell_cinema.dto.bookingDto;
 
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
@@ -15,6 +16,6 @@ public record BookingResponse(
         LocalTime time,
         Long movieId,
         boolean privateSpeaker,
-        double totalPriceSEK,
-        double totalPriceUSD) {
+        BigDecimal totalPriceSEK,
+        BigDecimal totalPriceUSD) {
 }
